Update about_variable_fonts.html
Spotted some rogue TK text (“[needs an illustration here]”) that needed to be cut.

Also, I propose moving the RWT.io figure down to the end of the related “Why this matters” subsection. The illustration is currently too distant from the text that discusses it imho—plus, visually the page gets awfully busy with the illustration _and_ the stacked demos in the right column.